Moved PyTest to testcommand.py
To minimize redundancy and improve style
