+++ python/setup.py
@@ -6,25 +6,10 @@ import sys
 from setuptools import setup
 from jsbeautifier.__version__ import __version__
 
-from setuptools.command.test import test as TestCommand
+import testcommand
+from testcommand import PyTest
 
-DIR = 'jsbeautifier/tests/'
-
-
-class PyTest(TestCommand):
-    user_options = [('pytest-args=', 'a', "Arguments to pass to py.test")]
-
-    def initialize_options(self):
-        TestCommand.initialize_options(self)
-        self.pytest_args = ['--assert=plain'] + [DIR + \
-            x for x in os.listdir(DIR) if x.endswith('.py') and x[0] not in '._']
-
-    def run_tests(self):
-        # import here, cause outside the eggs aren't loaded
-        import pytest
-        errno = pytest.main(self.pytest_args)
-        sys.exit(errno)
-
+testcommand.DIR = 'jsbeautifier/tests/'
 
 setup(name='jsbeautifier',
       version=__version__,
